requirements.txt
pad4pi==1.0.0
charlcd==0.4.0
RPi.GPIO==0.6.3